refactor: DAO 에 있던 Connection get, close 를 별도의 객체로 분리
