+++ src/main/java/chess/dbconnect/Connector.java
@@ -1,0 +1,8 @@
+package chess.dbconnect;
+
+import java.sql.Connection;
+
+public interface Connector {
+    Connection getConnection();
+    void closeConnection();
+}
